MPD-10323: pr review
