+++ meili_flutter/lib/meili_flutter.dart
@@ -3,4 +3,4 @@ export 'src/widgets/connect.dart';
 export 'src/widgets/meili_view.dart';
 
 export 'package:meili_flutter_platform_interface/meili_flutter_platform_interface.dart'
-    show MeiliParams, AvailParams, AdditionalParams, FlowType, MeiliFlutterPlatform;
+    show MeiliParams, AvailParams, AdditionalParams, BookingParams, FlowType, MeiliFlutterPlatform;
