+++ meili_flutter_platform_interface/lib/meili_flutter_platform_interface.dart
@@ -1,6 +1,6 @@
 export 'src/meili_flutter_platform_interface.dart';
 export 'src/method_channel_meili_flutter.dart';
+export 'src/model/additional_params.dart';
 export 'src/model/avail_params.dart';
-export 'src/model/booking_params.dart';
 export 'src/model/flow_enum.dart';
 export 'src/model/meili_params.dart';
